'ctrl+t' to add new paper!
