+++ src/containers/Paper/PaperAdd/PaperAdd.scss
@@ -1,0 +1,6 @@
+.paperAdd {
+  bottom: 70px;
+  padding: 0px 10% 0px 10%;
+  position: absolute;
+  width: 80%;
+}
