+++ src/redux/modules/hotkey.js
@@ -9,6 +9,7 @@ const initialState = new Map({
     destroyCard: 'd d',
     focusNewCardPanel: 'ctrl+n',
     focusCardListPanel: ['esc', 'ctrl+l'],
+    focusPaperAddPanel: ['ctrl+t'],
     forwardCardUp: 'k',
     forwardCardDown: 'j',
   },
